Moved LineChart.tsx's Interfaces
